requirements.txt
aiodns==3.2.0
aiofiles==0.8.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
appdirs==1.4.4
APScheduler==3.10.4
async-timeout==4.0.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.7.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langdetect==1.0.9
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
oauthlib==3.2.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
praw==7.7.1
prawcore==2.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
pyparsing==3.1.2
pyppeteer==2.0.0
PySocks==1.7.1
pytest==8.2.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.12.2
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.19
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
